refactor: rename BpmnGenerator
