+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/randomized/NewProcessExecutionRandomizePropertyTest.java
@@ -7,8 +7,8 @@
  */
 package io.camunda.zeebe.engine.processing.randomized;
 
-import io.camunda.process.generator.BpmnGenerator;
-import io.camunda.process.generator.BpmnGenerator.GeneratedProcess;
+import io.camunda.process.generator.BpmnProcessGenerator;
+import io.camunda.process.generator.BpmnProcessGenerator.GeneratedProcess;
 import io.camunda.process.generator.execution.ProcessExecutionStep;
 import io.camunda.zeebe.engine.util.EngineRule;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
@@ -29,7 +29,7 @@ import org.junit.runners.Parameterized.Parameters;
 public class NewProcessExecutionRandomizePropertyTest {
 
   public static final int AMOUNT_OF_PROCESSES = 1000;
-  private static final BpmnGenerator BPMN_GENERATOR = new BpmnGenerator();
+  private static final BpmnProcessGenerator BPMN_GENERATOR = new BpmnProcessGenerator();
   @Rule public final EngineRule engineRule = EngineRule.singlePartition();
   @Parameter public GeneratedProcess generatedProcess;
